refactor: 인증번호 관련 함수는 인증번호 컴포넌트에서 정의되도록 수정
